requirements.txt
opencv-python
flask
pytorch
numpy
werkzeug